refactor(internal): refactor local presets
