+++ lib/config/presets/local/index.ts
@@ -1,9 +1,9 @@
 import {
   PLATFORM_TYPE_GITHUB,
   PLATFORM_TYPE_GITLAB,
-} from '../../constants/platforms';
-import { RenovateConfig } from '../common';
-import { Preset } from './common';
+} from '../../../constants/platforms';
+import { RenovateConfig } from '../../common';
+import { Preset } from '../common';
 import * as github from './github';
 import * as gitlab from './gitlab';
